chore: extract types in their own files
