+++ src/types/Wallet.ts
@@ -1,0 +1,6 @@
+import type { Adapter, WalletReadyState } from "@solana/wallet-adapter-base";
+
+export type Wallet = {
+  adapter: Adapter;
+  readyState: WalletReadyState;
+};
