+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./Wallet";
+export * from "./WalletStore";
+export * from "./WalletStoreProps";
